+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/amigo/activities/NewPartyActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.util.Log
+import android.widget.Toast
 import com.amigotrip.amigo.R
 import com.amigotrip.amigo.datas.Party
 import com.amigotrip.amigo.remote.AmigoService
@@ -11,6 +12,7 @@ import kotlinx.android.synthetic.main.activity_new_party.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import java.util.regex.Pattern
 
 class NewPartyActivity : AppCompatActivity() {
 
@@ -39,16 +41,28 @@ class NewPartyActivity : AppCompatActivity() {
         }
 
 
+
+
         btn_submit.setOnClickListener {
 
             val name = input_name.text.toString()
-            val age = input_age.text.toString().toInt()
-            val language = input_lang.text.toString()
-            val theme = input_theme.text.toString()
+            //val age = input_age.text.toString().toInt()
+
+            val email = input_email.text.toString()
+
+            if (isEmailValid(email)) {
+                Toast.makeText(this, "apply done", Toast.LENGTH_SHORT).show()
+                startActivity(Intent(MainActivity@ this, NewPartyActivity::class.java))
+            } else {
+                input_email.error = "wrong email!"
+            }
+
+            val language = input_theme.text.toString()
+            val theme = input_attraction.text.toString()
             val attraction = input_attraction.text.toString()
 
             val party =
-                    Party(name, "wlals822@naver.com", age, gender, language, "12/24", theme,
+                    Party(name, "wlals822@naver.com", 20, gender, language, "12/24", theme,
                             attraction)
             val call = amigoService.newParty(party)
 
@@ -68,6 +82,15 @@ class NewPartyActivity : AppCompatActivity() {
 
             startActivity(Intent(NewPartyActivity@ this, DoneActivity::class.java))
         }
+    }
+
+    fun isEmailValid(email: String): Boolean {
+
+        val expression = "^[\\w\\.-]+@([\\w\\-]+\\.)+[A-Z]{2,4}$"
+        val pattern = Pattern.compile(expression, Pattern.CASE_INSENSITIVE)
+        val matcher = pattern.matcher(email)
+
+        return matcher.matches()
     }
 
     override fun onSupportNavigateUp(): Boolean {
